requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
Pillow==7.2.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.25.10
APScheduler==3.0.0
